Logic refactoring
